// src/types/api.ts
export interface components {
    schemas: {
// ...
        schemaReference: {
            /**
             * Format: uri
             * @description Schema ID
             */
            id: string;
            /** @description Schema name */
            name: string;
            /** @description Nested child schemas (recursive) */
            children?: components["schemas"]["schemaReference"][];
            /** @description Nested parent schemas (recursive) */
            parents?: components["schemas"]["schemaReference"][];
        };
// ...
    };
// ...
}